+++ lib/broker/greedy.go
@@ -3,6 +3,7 @@ package broker
 import (
 	"time"
 
+	"github.com/Jeffail/benthos/v3/internal/component/output"
 	"github.com/Jeffail/benthos/v3/lib/types"
 )
 
@@ -43,6 +44,19 @@ func (g *Greedy) Connected() bool {
 		}
 	}
 	return true
+}
+
+// MaxInFlight returns the maximum number of in flight messages permitted by the
+// output. This value can be used to determine a sensible value for parent
+// outputs, but should not be relied upon as part of dispatcher logic.
+func (g *Greedy) MaxInFlight() (m int, ok bool) {
+	for _, out := range g.outputs {
+		if mif, exists := output.GetMaxInFlight(out); exists && mif > m {
+			m = mif
+			ok = true
+		}
+	}
+	return
 }
 
 //------------------------------------------------------------------------------