+++ lib/output/broker.go
@@ -4,6 +4,7 @@ import (
 	"errors"
 	"fmt"
 
+	"github.com/Jeffail/benthos/v3/internal/component/output"
 	"github.com/Jeffail/benthos/v3/internal/docs"
 	"github.com/Jeffail/benthos/v3/internal/interop"
 	"github.com/Jeffail/benthos/v3/lib/broker"
@@ -93,9 +94,9 @@ outputs.`,
 			docs.FieldCommon("pattern", "The brokering pattern to use.").HasOptions(
 				"fan_out", "fan_out_sequential", "round_robin", "greedy",
 			),
-			docs.FieldCommon(
+			docs.FieldAdvanced(
 				"max_in_flight",
-				"The maximum number of messages to dispatch at any given time. Only relevant for `fan_out`, `fan_out_sequential` brokers.",
+				"The maximum number of parallel message batches to have in flight at any given time. Note that if a child output has a higher `max_in_flight` then the switch output will automatically match it, therefore this value is the minimum `max_in_flight` to set in cases where the child values can't be inferred (such as when using resource outputs as children). Only relevant for `fan_out`, `fan_out_sequential` brokers.",
 			),
 			docs.FieldCommon("outputs", "A list of child outputs to broker.").Array().HasType(docs.FieldOutput),
 			batch.FieldSpec(),
@@ -181,17 +182,24 @@ func NewBroker(
 		}
 	}
 
+	maxInFlight := conf.Broker.MaxInFlight
+	for _, out := range outputs {
+		if mif, ok := output.GetMaxInFlight(out); ok && mif > maxInFlight {
+			maxInFlight = mif
+		}
+	}
+
 	var b Type
 	switch conf.Broker.Pattern {
 	case "fan_out":
 		var bTmp *broker.FanOut
 		if bTmp, err = broker.NewFanOut(outputs, log, stats); err == nil {
-			b = bTmp.WithMaxInFlight(conf.Broker.MaxInFlight)
+			b = bTmp.WithMaxInFlight(maxInFlight)
 		}
 	case "fan_out_sequential":
 		var bTmp *broker.FanOutSequential
 		if bTmp, err = broker.NewFanOutSequential(outputs, log, stats); err == nil {
-			b = bTmp.WithMaxInFlight(conf.Broker.MaxInFlight)
+			b = bTmp.WithMaxInFlight(maxInFlight)
 		}
 	case "round_robin":
 		b, err = broker.NewRoundRobin(outputs, stats)