+++ lib/output/drop_on.go
@@ -7,6 +7,7 @@ import (
 	"fmt"
 	"time"
 
+	"github.com/Jeffail/benthos/v3/internal/component/output"
 	"github.com/Jeffail/benthos/v3/internal/docs"
 	"github.com/Jeffail/benthos/v3/lib/log"
 	"github.com/Jeffail/benthos/v3/lib/metrics"
@@ -308,6 +309,10 @@ func (d *dropOn) Consume(ts <-chan types.Transaction) error {
 // connected to its target.
 func (d *dropOn) Connected() bool {
 	return d.wrapped.Connected()
+}
+
+func (d *dropOn) MaxInFlight() (int, bool) {
+	return output.GetMaxInFlight(d.wrapped)
 }
 
 // CloseAsync shuts down the DropOnError input and stops processing requests.