+++ lib/output/drop_on_error.go
@@ -7,6 +7,7 @@ import (
 	"sync/atomic"
 	"time"
 
+	"github.com/Jeffail/benthos/v3/internal/component/output"
 	"github.com/Jeffail/benthos/v3/internal/docs"
 	"github.com/Jeffail/benthos/v3/lib/log"
 	"github.com/Jeffail/benthos/v3/lib/metrics"
@@ -227,6 +228,13 @@ func (d *DropOnError) Consume(ts <-chan types.Transaction) error {
 // connected to its target.
 func (d *DropOnError) Connected() bool {
 	return d.wrapped.Connected()
+}
+
+// MaxInFlight returns the maximum number of in flight messages permitted by the
+// output. This value can be used to determine a sensible value for parent
+// outputs, but should not be relied upon as part of dispatcher logic.
+func (d *DropOnError) MaxInFlight() (int, bool) {
+	return output.GetMaxInFlight(d.wrapped)
 }
 
 // CloseAsync shuts down the DropOnError input and stops processing requests.