+++ lib/output/retry.go
@@ -8,6 +8,7 @@ import (
 	"sync/atomic"
 	"time"
 
+	"github.com/Jeffail/benthos/v3/internal/component/output"
 	"github.com/Jeffail/benthos/v3/internal/docs"
 	"github.com/Jeffail/benthos/v3/lib/log"
 	"github.com/Jeffail/benthos/v3/lib/metrics"
@@ -309,6 +310,13 @@ func (r *Retry) Consume(ts <-chan types.Transaction) error {
 // connected to its target.
 func (r *Retry) Connected() bool {
 	return r.wrapped.Connected()
+}
+
+// MaxInFlight returns the maximum number of in flight messages permitted by the
+// output. This value can be used to determine a sensible value for parent
+// outputs, but should not be relied upon as part of dispatcher logic.
+func (r *Retry) MaxInFlight() (int, bool) {
+	return output.GetMaxInFlight(r.wrapped)
 }
 
 // CloseAsync shuts down the Retry input and stops processing requests.